+++ doc/version_history.rst
@@ -19,6 +19,8 @@ v0.7.0
 * Fix API docs.
 * Enable mypy type checking.
 * Change ``master`` to ``main`` in CONFIG_SCHEMA's ``id``, in preparation for renaming the branch.
+* Remove START and STOP commands from RPi Data Client.
+* The sensor name, timestamp, response code and data are encoded as separate named entities.
 
 Requires: